doc - restructure scripts, separate moods from entries
