+++ scripts/journalEntries/JournalDataProvider.js
