+++ scripts/journalEntries/JournalEntry.js
