+++ scripts/main.js
@@ -1,6 +1,6 @@
-import { getEntries } from "./JournalDataProvider.js";
-import { EntryListComponent } from "./JournalEntryList.js"
-import { JournalFormComponent } from "./JournalFormComponent.js"
+import { getEntries } from "./journalEntries/JournalDataProvider.js";
+import { EntryListComponent } from "./journalEntries/JournalEntryList.js"
+import { JournalFormComponent } from "./journalEntries/JournalFormComponent.js"
 
 
 JournalFormComponent()
